requirements.txt
agent-framework[all]
python-dotenv
mcp[cli]
PyJWT[crypto]
httpx
